end of split WebSocket to external file
